+++ packages/cloud_firestore/cloud_firestore/lib/src/query.dart
@@ -41,7 +41,8 @@ class Query {
     return (operator == '<' ||
         operator == '<=' ||
         operator == '>' ||
-        operator == '>=');
+        operator == '>=' ||
+        operator == '!=');
   }
 
   /// Asserts that a [DocumentSnapshot] can be used within the current
@@ -339,6 +340,7 @@ class Query {
   Query where(
     dynamic field, {
     dynamic isEqualTo,
+    dynamic isNotEqualTo,
     dynamic isLessThan,
     dynamic isLessThanOrEqualTo,
     dynamic isGreaterThan,
@@ -346,6 +348,7 @@ class Query {
     dynamic arrayContains,
     List<dynamic> arrayContainsAny,
     List<dynamic> whereIn,
+    List<dynamic> whereNotIn,
     bool isNull,
   }) {
     _assertValidFieldType(field);
@@ -376,6 +379,7 @@ class Query {
     }
 
     if (isEqualTo != null) addCondition(field, '==', isEqualTo);
+    if (isNotEqualTo != null) addCondition(field, '!=', isNotEqualTo);
     if (isLessThan != null) addCondition(field, '<', isLessThan);
     if (isLessThanOrEqualTo != null) {
       addCondition(field, '<=', isLessThanOrEqualTo);
@@ -391,6 +395,7 @@ class Query {
       addCondition(field, 'array-contains-any', arrayContainsAny);
     }
     if (whereIn != null) addCondition(field, 'in', whereIn);
+    if (whereNotIn != null) addCondition(field, 'not-in', whereNotIn);
     if (isNull != null) {
       assert(
           isNull,
@@ -401,6 +406,8 @@ class Query {
 
     dynamic hasInequality;
     bool hasIn = false;
+    bool hasNotIn = false;
+    bool hasNotEqualTo = false;
     bool hasArrayContains = false;
     bool hasArrayContainsAny = false;
 
@@ -424,7 +431,9 @@ class Query {
             'You can only perform equals comparisons on null.');
       }
 
-      if (operator == 'in' || operator == 'array-contains-any') {
+      if (operator == 'in' ||
+          operator == 'array-contains-any' ||
+          operator == 'not-in') {
         assert(value is List,
             "A non-empty [List] is required for '$operator' filters.");
         assert((value as List).length <= 10,
@@ -435,8 +444,20 @@ class Query {
             "'$operator' filters cannot contain 'null' in the [List].");
       }
 
+      if (operator == '!=') {
+        assert(!hasNotEqualTo, "You cannot use '!=' filters more than once.");
+        assert(!hasNotIn, "You cannot use '!=' filters with 'not-in' filters.");
+        hasNotEqualTo = true;
+      }
+
+      if (operator == 'not-in') {
+        assert(!hasNotIn, "You cannot use 'not-in' filters more than once.");
+        assert(!hasNotEqualTo,
+            "You cannot use 'not-in' filters with '!=' filters.");
+      }
+
       if (operator == 'in') {
-        assert(!hasIn, "You cannot use 'in' filters more than once.");
+        assert(!hasIn, "You cannot use 'whereIn' filters more than once.");
         hasIn = true;
       }