+++ packages/firebase_auth/firebase_auth_web/lib/src/interop/firebase_interop.dart
@@ -1,0 +1,16 @@
+// Copyright 2020, the Chromium project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// ignore_for_file: public_member_api_docs
+
+@JS('firebase')
+library firebase_interop.firebase;
+
+import 'package:firebase_core_web/firebase_core_web_interop.dart';
+
+import 'package:js/js.dart';
+import 'auth_interop.dart';
+
+@JS()
+external AuthJsImpl auth([AppJsImpl app]);