+++ packages/cloud_firestore/cloud_firestore/CHANGELOG.md
@@ -1,3 +1,46 @@
+## 0.14.3
+
+ - **FEAT**: migrate firebase interop files to local repository (#3973).
+ - **FEAT**: add not-in & != query support (#3748).
+ - **FEAT**: bump `compileSdkVersion` to 29 in preparation for upcoming Play Store requirement.
+ - **FEAT** [WEB] `FirebaseFirestore.enablePersistence` now accepts `PersistenceSettings`
+ - **FEAT** [WEB] adds `PersistenceSettings` class
+ - **FEAT** [WEB] adds support for `FirebaseFirestore.clearPersistence`
+ - **FEAT** [WEB] adds support for `FirebaseFirestore.terminate`
+ - **FEAT** [WEB] adds support for `FirebaseFirestore.waitForPendingWrites`
+ - **FEAT** [WEB] adds support for `SetOptions.mergeFields`
+ - **FEAT** [WEB] adds `GetOptions` support for querying against server/cache
+ - **FEAT** [WEB] adds support for `Query.limitToLast`
+ - **FEAT** [WEB] adds support for `FirebaseFirestore.snapshotsInSync`
+
+## 0.14.2
+
+ - **FEAT**: bump compileSdkVersion to 29 (#3975).
+ - **FEAT**: update Firebase iOS SDK version to 6.33.0 (from 6.26.0).
+ - **CHORE**: update Firestore example app podfile.
+
+## 0.14.1+3
+
+ - **FIX**: remove unused dart:async import (#3611).
+ - **FIX**: fix returning of transaction result (#3747).
+
+## 0.14.1+2
+
+ - Update a dependency to the latest release.
+
+## 0.14.1+1
+
+ - **FIX**: remove listener if available (#3452).
+ - **DOCS**: remove `updateBlock` reference in Firestore docs (#3728).
+
+## 0.14.1
+
+ - **FIX**: local dependencies in example apps (#3319).
+ - **FIX**: pub.dev score fixes (#3318).
+ - **FIX**: add missing deprecated static methods (#3278).
+ - **FEAT**: add a [] operator to DocumentSnapshot, acting as get() (#3387).
+ - **DOCS**: Fixed docs typo (#3471).
+
 ## 0.14.0+2
 
 * Added missing deprecated `Firestore` static methods.