+++ packages/firebase_database/CHANGELOG.md
@@ -1,3 +1,25 @@
+## 4.3.0
+
+ - **FEAT**: bump compileSdkVersion to 29 (#3975).
+ - **FEAT**: bump `compileSdkVersion` to 29 in preparation for upcoming Play Store requirement.
+ - **CHORE**: publish packages.
+ - **CHORE**: publish packages.
+
+## 4.2.0
+
+ - **FEAT**: bump compileSdkVersion to 29 (#3975).
+ - **FEAT**: update Firebase iOS SDK version to 6.33.0 (from 6.26.0).
+
+## 4.1.1
+
+ - Update a dependency to the latest release.
+
+## 4.1.0
+
+ - **FIX**: local dependencies in example apps (#3319).
+ - **FEAT**: implement support for `ServerValue.increment(delta)` (#3109).
+ - **CHORE**: intellij cleanup (#3326).
+
 ## 4.0.0
 
 * Add MacOS support.